+++ tests/integration/cases/scalar_bool/Rust_declaration_style_lazy_static.rs
@@ -1,0 +1,5 @@
+use std::sync::LazyLock;
+fn main() {
+    static my_data: LazyLock<bool> = LazyLock::new(|| true);
+    let _ = my_data;
+}