+++ tests/integration/cases/scalar_string/Rust_declaration_style_lazy_static.rs
@@ -1,0 +1,5 @@
+use std::sync::LazyLock;
+fn main() {
+    static my_data: LazyLock<&str> = LazyLock::new(|| "hello");
+    let _ = my_data;
+}